Refactor polars statistics
